BMW/requirement.txt
pillow
pytesseract
cent
django-countries
reportlab
xhtml2pdf
pytz
phonenumbers
twilio==5.6.0
paypalrestsdk
oauth2client==2.1.0
icalendar
google-api-python-client==1.4.0
pdfcrowd
pytesseract
html5lib==1.0b8